backend/requirements.txt
# Core dependencies
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
python-multipart==0.0.6
requests==2.32.3

# Whisper + OpenAI
openai==0.28.0
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
ffmpeg-python==0.2.0
torch==2.8.0
numpy==1.26.4

# Pydantic / Validation
pydantic==2.11.7
pydantic-core==2.33.2
typing-inspection==0.4.1
typing_extensions==4.14.1
annotated-types==0.7.0

# Optional: Document processing (used by your app)
python-docx==1.1.2
PyPDF2==3.0.1
python-pptx==1.0.2

# Utility + encoding
tiktoken==0.9.0
protobuf==4.25.8
tenacity==8.5.0

# Optional: Database
psycopg2-binary==2.9.9

# Optional: HTML & file handling
Jinja2==3.1.6
MarkupSafe==3.0.2
